Updated code directory to build correctly
The solutions directory had already been updated to use the correct
libraries and the updated path and packages (com.allendowney.thinkdast),
but there was no build.xml in the code directory, and the one in the
(obsolete?) lab01 directory used the old path and package
(com.flatironschool.javacs).  This commit adds an updated build.xml to
the code directory and removes the lab01 directory.